refactor : 차 이름 원시값 포장
